src/requirements.txt
asyncpg==0.29.0
fastapi==0.111.1
fastapi-cli==0.0.4
httpx==0.27.0
python-dotenv==1.0.1
boto3==1.34.136
pytz==2024.1